test: cleanup BDD tests
Closes #22
