--- test/bdd/fixtures/oathkeeper/rules/resource-server.json
@@ -1,34 +1,0 @@
-[
-  {
-    "id": "edv-resource-server-rule",
-    "upstream": {
-      "url": "https://edv.example.com:8080"
-    },
-    "match": {
-      "url": "http://localhost:4455/encrypted-data-vaults<.*>",
-      "methods": [
-        "GET",
-        "POST",
-        "DELETE"
-      ]
-    },
-    "authenticators": [
-      {
-        "handler": "oauth2_introspection"
-      }
-    ],
-    "mutators": [
-      {
-        "handler": "header",
-        "config": {
-          "headers": {
-            "X-User": "{{ print .Subject }}"
-          }
-        }
-      }
-    ],
-    "authorizer": {
-      "handler": "allow"
-    }
-  }
-]
